+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/langchain
 
+## 3.0.0-beta.20
+
+### Patch Changes
+
+- ai@7.0.0-beta.20
+
+## 3.0.0-beta.19
+
+### Patch Changes
+
+- ai@7.0.0-beta.19
+
+## 3.0.0-beta.18
+
+### Patch Changes
+
+- ai@7.0.0-beta.18
+
 ## 3.0.0-beta.17
 
 ### Patch Changes