+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,24 @@
 
 ### Patch Changes
 
+- ai@7.0.0-beta.20
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@7.0.0-beta.19
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@7.0.0-beta.18
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@7.0.0-beta.17
 
 ## 0.0.1-beta.0