+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/svelte
 
+## 5.0.0-beta.20
+
+### Patch Changes
+
+- ai@7.0.0-beta.20
+
+## 5.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies [f7d4f01]
+  - @ai-sdk/provider-utils@5.0.0-beta.3
+  - ai@7.0.0-beta.19
+
+## 5.0.0-beta.18
+
+### Patch Changes
+
+- ai@7.0.0-beta.18
+- @ai-sdk/provider-utils@5.0.0-beta.2
+
 ## 5.0.0-beta.17
 
 ### Patch Changes